Feat(Authentication) Added registration

+++ studybud/base/views.py
@@ -7,7 +7,8 @@ from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib import messages
 from django.contrib.auth import authenticate,login,logout
-
+from django.contrib.auth.decorators import login_required
+from django.http  import HttpResponse
 # Create your views here.
 
 # rooms=[
@@ -17,6 +18,10 @@ from django.contrib.auth import authenticate,login,logout
 # ]
 
 def loginPage(request):
+    page='login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
 
     if request.method =="POST":
         username=request.POST.get('username')
@@ -37,12 +42,17 @@ def loginPage(request):
             messages.error(request,'Username or Password does not exists')
         
 
-    context={}
+    context={'page':page}
     return render(request,'base/login_register.html',context)
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
+
+def registerPage(request):
+    page='register'
+    return render(request,'base/login_register.html')
 
 def home(request):
 
@@ -67,6 +77,7 @@ def room(request,pk):
     return render(request,'base/room.html',context)
 
 
+@login_required(login_url='login')
 def createRoom(request):
     form=RoomForm()
     
@@ -82,10 +93,16 @@ def createRoom(request):
     return render(request,'base/room_form.html',context)
 
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
 
     room=Room.objects.get(id=pk)
     form=RoomForm(instance=room)
+
+    if request.user!=room.host:
+        return HttpResponse('You are not allowed to perform this action!')
+        messages.error(request,'Sorry only the room creator can delete a room')
+        return redirect('home')
 
     if request.method=="POST":
         form=RoomForm(request.POST,instance=room)
@@ -96,9 +113,17 @@ def updateRoom(request,pk):
     context={'form':form}
     return render(request,'base/room_form.html',context)
 
-
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room=Room.objects.get(id=pk)
+
+    if request.user!=room.host:
+        return HttpResponse('You are not allowed to perform this action!')
+        messages.error(request,'Sorry only the room creator can delete a room')
+        return redirect('home')
+
+    
+
     if request.method=="POST":
         room.delete()
         return redirect('home')
